bash:  fix prompt corruption
After exiting the curses application, the command line was getting
printed one line to high and the correct line was not being cleared.
This made it pretty are to figure out what had happened.

By changing the macro to yank the command line, hit enter then paste the
command line, the corruption seems to be fixed.