scanner:  add a post-match transform
It is typically convenient to present simpler matches to the user which
need to be transformed into the full match later on.  For example, when
scanning a git directory, it's nice to see paths rooted from the top of
the clone rather than the full or relative path.  Support this by adding
a transform method to scanners.